+++ out/test/extension.test.js
@@ -22,41 +22,69 @@ var __importStar = (this && this.__importStar) || function (mod) {
     __setModuleDefault(result, mod);
     return result;
 };
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
 Object.defineProperty(exports, "__esModule", { value: true });
 const assert = __importStar(require("assert"));
-const MockTypes = __importStar(require("./test"));
-const parse_1 = __importDefault(require("../utils/parse"));
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
 const vscode = __importStar(require("vscode"));
-// import * as myExtension from '../../extension';
+const MockTypes = __importStar(require("./test"));
+const processor_1 = require("../utils/singleType/processor");
+const processor_2 = require("../utils/complexTypes/processor");
 suite("Extension Test Suite", () => {
     vscode.window.showInformationMessage("Start all tests.");
+    // Single type tests
     test("Simple object", () => {
-        const result = (0, parse_1.default)(MockTypes.Test1);
+        const result = (0, processor_1.singleTypeProcessor)(MockTypes.Test1);
         const parse = JSON.parse(result);
         assert.deepEqual({ a: "", b: 0, c: [], d: "undefined" }, parse);
     });
     test("Simple object with comments", () => {
-        const result = (0, parse_1.default)(MockTypes.Test2);
+        const result = (0, processor_1.singleTypeProcessor)(MockTypes.Test2);
         const parse = JSON.parse(result);
         assert.deepEqual({ a: "", b: 0, c: [], d: "undefined" }, parse);
     });
     test("Complicating object with comments", () => {
-        const result = (0, parse_1.default)(MockTypes.Test3);
+        const result = (0, processor_1.singleTypeProcessor)(MockTypes.Test3);
+        console.log("result", result);
         const parse = JSON.parse(result);
         assert.deepEqual({
             adsf: "",
             b: 0,
             cvcxv: [],
+            "123": "any",
+            z123: null,
             d: "undefined",
             e: {},
             f: { t: { a: "", b: [], c: { h: { k: "" } } } },
             k: {},
             o: { t: { a: "", b: [], c: { h: { j: "", n: 0 } } } },
+        }, parse);
+    });
+    // Multi types tests
+    test("Multiple types with a root type", () => {
+        const result = (0, processor_2.complexTypesProcessor)(MockTypes.Test4);
+        const parse = JSON.parse(result);
+        assert.deepEqual({
+            Child1: { esfdsa: "", vcxzvcxz: "", tewtwe: "", qqew: "" },
+            Child2: {
+                dasda: "",
+                gfxzgf: "",
+                bcvk4: "",
+                wqe123: "",
+                kdkeiII: "",
+            },
+            Info: {
+                FEFEss: 0,
+                fskeo: 0,
+                tfkr4D: 0,
+                fddsfe: "",
+                azfds: { esfdsa: "", vcxzvcxz: "", tewtwe: "", qqew: "" },
+                bvnb: "",
+                tests: [],
+                qqqqaaaaa: [
+                    { dasda: "", gfxzgf: "", bcvk4: "", wqe123: "", kdkeiII: "" },
+                ],
+                ttt: [{ dasda: "", gfxzgf: "", bcvk4: "", wqe123: "", kdkeiII: "" }],
+                ccccxxxxx: [{ esfdsa: "", vcxzvcxz: "", tewtwe: "", qqew: "" }],
+            },
         }, parse);
     });
 });